rss-merger/requirements.txt
requests==2.28.2
lxml==4.9.2
tinytag==1.8.1
yt_dlp==2023.1.6
python_dateutil==2.8.2
PyYAML==6.0
